+++ tests/integration_tests/hardhat/contracts/Greeter.sol
@@ -1,6 +1,7 @@
 pragma solidity >0.5.0;
 
 contract Greeter {
+    uint public n;
     string public greeting;
 
     event ChangeGreeting(address from, string value);
@@ -16,5 +17,9 @@ contract Greeter {
 
     function greet() public view returns (string memory) {
         return greeting;
+    }
+
+    function intValue() public view returns (uint) {
+        return n;
     }
 }