+++ x/evm/statedb/interfaces.go
@@ -32,6 +32,7 @@ type Keeper interface {
 
 	AddBalance(ctx sdk.Context, addr sdk.AccAddress, coins sdk.Coins) error
 	SubBalance(ctx sdk.Context, addr sdk.AccAddress, coins sdk.Coins) error
+	SetBalance(ctx sdk.Context, addr common.Address, amount *big.Int) error
 	GetBalance(ctx sdk.Context, addr sdk.AccAddress, denom string) *big.Int
 
 	// Read methods